Rename DashboardJoinRequests to List
