+++ src/components/DashboardJoinList/index.js
@@ -1,0 +1,9 @@
+import DashboardJoinListContainer from 'components/DashboardJoinList/DashboardJoinListContainer'
+import DashboardJoinListData from 'components/DashboardJoinList/DashboardJoinListData'
+import DashboardJoinListPresentation from 'components/DashboardJoinList/DashboardJoinListPresentation'
+
+export default {
+  Container: DashboardJoinListContainer,
+  Data: DashboardJoinListData,
+  Presentation: DashboardJoinListPresentation,
+}
