--- src/components/DashboardJoinRequests/index.js
@@ -1,9 +1,0 @@
-import DashboardJoinRequestsContainer from 'components/DashboardJoinRequests/DashboardJoinRequestsContainer'
-import DashboardJoinRequestsData from 'components/DashboardJoinRequests/DashboardJoinRequestsData'
-import DashboardJoinRequestsPresentation from 'components/DashboardJoinRequests/DashboardJoinRequestsPresentation'
-
-export default {
-  Container: DashboardJoinRequestsContainer,
-  Data: DashboardJoinRequestsData,
-  Presentation: DashboardJoinRequestsPresentation,
-}
